# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aaryajanani Content Ops & Project Memory Platform API
  version: 1.0.0
  description: |
    Backend API for the Content Operations & Project Memory Platform (Aaryajanani).
    Node.js / Express / MongoDB.

    ## Authentication
    - Register or login to obtain an `accessToken` (JWT, 15 min) and a `refreshToken`.
    - Send the access token as `Authorization: Bearer <token>` on protected endpoints.
    - When the access token expires, call `POST /auth/refresh` with the `refreshToken`
      (rotates the token; the old refresh token is revoked).
    - Logout revokes the current refresh token.
  contact:
    name: Aaryajanani Backend
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development (default apiPrefix /api/v1)
...
paths:
...
  /projects/{projectId}/brief:
    get:
      tags: [Briefs]
      summary: Get the project's structured brief
      operationId: getBrief
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Brief (exists:false if not created yet)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Brief'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Briefs]
      summary: Create or update the project brief
      operationId: upsertBrief
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BriefRequest'
      responses:
        '200':
          description: Brief saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Brief'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing brief.manage permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  # ================= Reviews =================
...
  /projects/{projectId}/channels:
    get:
      tags: [Chat]
      summary: List project channels
      operationId: listChannels
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Channels
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Channel'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Chat]
      summary: Create a channel (topic or DM)
      operationId: createChannel
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Channel'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing chat.post permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /projects/{projectId}/channels/{channelId}/messages:
    get:
      tags: [Chat]
      summary: List messages in a channel (optionally a thread)
      operationId: listMessages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ChannelId'
        - name: parentId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          description: Filter to a threaded reply (omit for top-level)
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Skip'
      responses:
        '200':
          description: Paginated messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ChatMessage'
                      total:
                        type: integer
                      limit:
                        type: integer
                      skip:
                        type: integer
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Chat]
      summary: Post a message (or threaded reply) in a channel
      operationId: postMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostMessageRequest'
      responses:
        '201':
          description: Message posted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ChatMessage'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing chat.post permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Channel or parent message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  # ================= Contracts =================
  /organizations/{organizationId}/contracts:
    get:
      tags: [Contracts]
      summary: List contracts for an organization
      operationId: listContracts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, sent, viewed, signed]
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Skip'
      responses:
        '200':
          description: Paginated contracts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Contract'
                      total:
                        type: integer
                      limit:
                        type: integer
                      skip:
                        type: integer
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing contract.manage permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Contracts]
      summary: Create a contract (draft)
      operationId: createContract
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractRequest'
      responses:
        '201':
          description: Contract created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Contract'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing contract.manage permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /organizations/{organizationId}/contracts/{contractId}:
    get:
      tags: [Contracts]
      summary: Get a contract
      operationId: getContract
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/ContractId'
      responses:
        '200':
          description: Contract found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Contract'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Contracts]
      summary: Edit a draft contract
      operationId: updateContract
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/ContractId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractRequest'
      responses:
        '200':
          description: Contract updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Contract'
        '400':
          description: Only draft contracts can be edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /organizations/{organizationId}/contracts/{contractId}/send:
    post:
      tags: [Contracts]
      summary: Send a contract for signature (draft -> sent)
      operationId: sendContract
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/ContractId'
      responses:
        '200':
          description: Contract sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Contract'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /organizations/{organizationId}/contracts/{contractId}/sign:
    post:
      tags: [Contracts]
      summary: Sign a contract (captures e-signature)
      operationId: signContract
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/ContractId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignContractRequest'
      responses:
        '200':
          description: Contract signed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Contract'
        '400':
          description: Contract already signed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  # ================= Invoices =================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned by /auth/login or /auth/refresh (15 min expiry)
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      description: MongoDB ObjectId of the organization
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      description: MongoDB ObjectId of the project
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Maximum number of items to return
    Skip:
      name: skip
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip (offset pagination)
...
    ChannelId:
      name: channelId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      description: MongoDB ObjectId of the chat channel
    ContractId:
      name: contractId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      description: MongoDB ObjectId of the contract
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              examples: [not_found, unauthorized, forbidden, conflict, internal_error]
            message:
              type: string
    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              const: validation_error
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # ---------- Tasks ----------
...
    BriefRequest:
      type: object
      additionalProperties: false
      properties:
        goal:
          type: string
          maxLength: 5000
          example: 'Launch promo video for Q3 product drop'
        targetAudience:
          type: string
          maxLength: 2000
        videoType:
          type: string
          maxLength: 200
        references:
          type: array
          items:
            type: string
          maxItems: 100
        deliverables:
          type: array
          items:
            type: string
          maxItems: 100
        deadline:
          type: string
          format: date
        brandGuidelines:
          type: string
          maxLength: 5000
        specialInstructions:
          type: string
          maxLength: 5000
    Brief:
      type: object
      properties:
        _id:
          type: string
        projectId:
          type: string
        goal:
          type: string
          nullable: true
        targetAudience:
          type: string
          nullable: true
        videoType:
          type: string
          nullable: true
        references:
          type: array
          items:
            type: string
        deliverables:
          type: array
          items:
            type: string
        deadline:
          type: string
          format: date
          nullable: true
        brandGuidelines:
          type: string
          nullable: true
        specialInstructions:
          type: string
          nullable: true
        submittedBy:
          type: string
        exists:
          type: boolean
          description: false when no brief has been created yet
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ---------- Reviews ----------
...
    CreateChannelRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: general
        type:
          type: string
          enum: [channel, dm]
          default: channel
        members:
          type: array
          items:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          description: Member user ids (for DMs)
    PostMessageRequest:
      type: object
      required: [body]
      additionalProperties: false
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 10000
        parentId:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          description: Parent message id to reply in a thread
        attachments:
          type: array
          items:
            type: string
          maxItems: 20
    Channel:
      type: object
      properties:
        _id:
          type: string
        projectId:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [channel, dm]
        members:
          type: array
          items:
            type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ChatMessage:
      type: object
      properties:
        _id:
          type: string
        channelId:
          type: string
        projectId:
          type: string
        author:
          type: string
        parentId:
          type: string
          nullable: true
        body:
          type: string
        attachments:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ---------- Contracts ----------
    ContractRequest:
      type: object
      additionalProperties: false
      properties:
        projectId:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: 'Video production agreement'
        terms:
          type: string
          maxLength: 20000
        amount:
          type: number
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
          default: USD
        sentTo:
          type: string
          description: Client name or email the contract is sent to
    SignContractRequest:
      type: object
      required: [signerName, signerEmail]
      additionalProperties: false
      properties:
        signerName:
          type: string
          minLength: 1
          maxLength: 200
        signerEmail:
          type: string
          format: email
    Contract:
      type: object
      properties:
        _id:
          type: string
        organizationId:
          type: string
        projectId:
          type: string
          nullable: true
        title:
          type: string
        terms:
          type: string
          nullable: true
        amount:
          type: number
          nullable: true
        currency:
          type: string
        status:
          type: string
          enum: [draft, sent, viewed, signed]
        sentTo:
          type: string
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
        viewedAt:
          type: string
          format: date-time
          nullable: true
        signedAt:
          type: string
          format: date-time
          nullable: true
        signedBy:
          type: string
          nullable: true
        signerName:
          type: string
          nullable: true
        signerEmail:
          type: string
          nullable: true
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ---------- Invoices ----------
...